+++ calres.php
@@ -19,16 +19,13 @@ $request = new Request(false);
 $path = $request->path();
 if (empty($path) || $path == '/') {
     Request::notFound();
-    return;
 }
 $path = substr($path, 1);
 if (!preg_match('/^\d+\/\w+\/\w+$/', $path)) {
     Request::notFound();
-    return;
 }
 [$database, $alg, $digest] = explode('/', $path);
 $hash = $alg . ':' . $digest;
-if (!Resource::sendImageResource($hash, null, $database)) {
+if (!Resource::sendImageResource($hash, null, intval($database))) {
     Request::notFound();
-    return;
 }